Op Names; Op graphviz node attributes
Ops implicitly have a name based on their classname, this can be less than usefule
when debugging. This commit allows users to pass a 'name' parameter on Op creation.
Currently, this friendly name is only used for DAG.viz() - due to the limitation that
caching requires the save name to be a valid python name.

Color and style for graphiv are broken out as method calls on the Op. Simply overide
the method to return a different value to change the way a new Op appears in the DAG