WIP removing metaclass; better dedup and better hashing
Use of didn't solve the problems it needed to, but added a bit of complexity.
This commit is a mojr step towards removing this aspect and thus simplifying the
code base and fixing a few bugs.

Metaclass is now only used ina few places that are not being used currently,
need to update tests before complete removal.