TQDM uses friendly name; downstream rerun; getitem completed ops
new DAG construction parameter 'force_downstream_rerun' will force
all ops that are downstream of an executed op (i.e. not cached) to be
rerun. This was the default behavior previously, but the addition of this
option allows the user to turn this feature off by setting this parameter
to False.

Item access on the DAG will get from the completed_ops dictionary